# storage/remote: speed up StoreSeries by re-using labels.Builder